Moved Request body parsing to Request.
