+++ src/main/java/com/djrapitops/permissionsex/backends/web/http/Request.java
@@ -1,12 +1,16 @@
 package com.djrapitops.permissionsex.backends.web.http;
 
 import com.djrapitops.permissionsex.backends.web.http.auth.Authentication;
+import com.djrapitops.permissionsex.exceptions.ParseException;
+import com.djrapitops.permissionsex.utilities.Closer;
 import com.djrapitops.permissionsex.utilities.Wrapper;
 import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpExchange;
 
+import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.util.zip.GZIPInputStream;
 
 /**
@@ -22,6 +26,8 @@ public class Request {
 
     private Wrapper<InputStream> requestBodyWrapper;
     private Headers requestHeaders;
+
+    private String requestBody;
 
     public Request(HttpExchange exchange) {
         requestMethod = exchange.getRequestMethod();
@@ -51,7 +57,46 @@ public class Request {
         return auth != null;
     }
 
-    public InputStream getRequestBody() throws IOException {
+    public String getRequestBodyString() throws ParseException {
+        if (requestBody == null) {
+            requestBody = getStringFromRequestBody();
+        }
+        return requestBody;
+    }
+
+    private String getStringFromRequestBody() throws ParseException {
+        try {
+            return readInputStream(getRequestBody());
+        } catch (IOException e) {
+            throw new ParseException("Failed to get request body.", e);
+        }
+    }
+
+    private String readInputStream(InputStream inputStream) throws ParseException {
+        StringBuilder builder = new StringBuilder();
+        BufferedReader bufferedReader = null;
+        try {
+            if (inputStream != null) {
+                bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
+                char[] charBuffer = new char[128];
+                int bytesRead = -1;
+                while ((bytesRead = bufferedReader.read(charBuffer)) > 0) {
+                    builder.append(charBuffer, 0, bytesRead);
+                }
+            } else {
+                builder.append("");
+            }
+        } catch (IOException e) {
+            throw new ParseException("Failed to read request body.", e);
+        } finally {
+            Closer.ignoreExceptions(bufferedReader);
+            Closer.ignoreExceptions(inputStream);
+        }
+
+        return builder.toString();
+    }
+
+    private InputStream getRequestBody() throws IOException {
         String contentEncoding = getRequestHeader("Content-Encoding");
         if ("gzip".equals(contentEncoding)) {
             return new GZIPInputStream(requestBodyWrapper.get());
