+++ src/main/java/com/djrapitops/permissionsex/backends/web/pages/RestAPIHandler.java
@@ -5,14 +5,8 @@ import com.djrapitops.permissionsex.backends.web.http.Response;
 import com.djrapitops.permissionsex.backends.web.http.auth.Authentication;
 import com.djrapitops.permissionsex.backends.web.http.responses.JsonErrorResponse;
 import com.djrapitops.permissionsex.exceptions.ParseException;
-import com.djrapitops.permissionsex.utilities.Closer;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 
 /**
  * Abstract TreePageHandler for all RestAPI providers.
@@ -34,36 +28,9 @@ public abstract class RestAPIHandler extends TreePageHandler {
         return null;
     }
 
+    @Deprecated
     protected String getStringFromRequestBody(Request request) throws ParseException {
-        try {
-            return readInputStream(request.getRequestBody());
-        } catch (IOException e) {
-            throw new ParseException("Failed to get request body.", e);
-        }
-    }
-
-    private String readInputStream(InputStream inputStream) throws ParseException {
-        StringBuilder builder = new StringBuilder();
-        BufferedReader bufferedReader = null;
-        try {
-            if (inputStream != null) {
-                bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
-                char[] charBuffer = new char[128];
-                int bytesRead = -1;
-                while ((bytesRead = bufferedReader.read(charBuffer)) > 0) {
-                    builder.append(charBuffer, 0, bytesRead);
-                }
-            } else {
-                builder.append("");
-            }
-        } catch (IOException e) {
-            throw new ParseException("Failed to read request body.", e);
-        } finally {
-            Closer.ignoreExceptions(bufferedReader);
-            Closer.ignoreExceptions(inputStream);
-        }
-
-        return builder.toString();
+        return request.getRequestBodyString();
     }
 
     /**
