+++ src/main/java/com/djrapitops/permissionsex/backends/web/pages/rest/WorldRestAPI.java
@@ -37,7 +37,9 @@ public class WorldRestAPI extends RestAPIHandler {
             if ("PUT".equals(requestMethod)) {
                 // PUT /api/worlds/ - updates worlds when "Save Changes" is pressed
                 try {
-                    worldJSONService.updateWorlds((JsonArray) parseJSONFromString(getStringFromRequestBody(request)));
+                    worldJSONService.updateWorlds((JsonArray) parseJSONFromString(request.getRequestBodyString()));
+                } catch (ClassCastException e) {
+                    return new JsonErrorResponse("Sent JSON was not an Array", 400);
                 } catch (ParseException e) {
                     return e.getCause() == null ?
                             new JsonErrorResponse(e.getMessage(), 500) :
